// src/support/enums.ts
export enum NAVIGATION_ITEMS_NAMES {
    TUTORIALS = 'Tutorials ',
    REFERENCES = 'References ',
    EXERCISES = 'Exercises ',
    VIDEOS = 'Videos'
}

export enum PAGES {
    HOME = 'HomePage',
    LOGIN = 'LoginPage',
    RESET_PASSWORD = 'ResetPasswordPage',
    SIGN_UP = 'SignUpPage'
}

export enum PASSWORD_VALIDATION_HELPERS {
    LOWERCASE_CHARACTER = 'One lowercase character',
    UPPERCASE_CHARACTER = 'One uppercase character',
    ONE_NUMBER = 'One number',
    ONE_SPECIAL_CHARACTER = 'One special character',
    MINIMUM_CHARACTERS = '8 characters minimum'
}

export enum SIGN_UP_NAME_TYPES {
    FIRST_NAME = 'first_name',
    LAST_NAME = 'last_name'
}

export enum SPECIAL_SYMBOLS {
    tilde = '~',
    exclamation = '!',
    hash = '#',
    percent = '%',
    caret = '^',
    asterisk = '*',
    hyphen = '-',
    equals = '=',
    plus = '+',
    semicolon = ';',
}